Enhances authentication and authorization
Improves security with JWT authentication and role-based authorization.
Adds admin role and authorization middleware.
Refactors authentication flow and fixes token handling.
Introduces tooltips for improved user experience.
Updates UI components and implements admin-specific routes.

+++ Reboost-ws/src/service/gebruiker.ts
@@ -34,7 +34,7 @@ export const checkAndParseSession = async (
     const { roles, sub } = await verifyJWT(authToken);
 
     return {
-      userId: Number(sub),
+      gebruikersId: Number(sub),
       roles,
     };
   } catch (error: any) {
@@ -116,7 +116,7 @@ export const getAll = async (): Promise<PublicGebruiker[]> => {
 
 export const getById = async (id: number): Promise<PublicGebruiker> => {
   const gebruiker = await prisma.gebruiker.findUnique({ where: { id } });
-
+  console.log('Fetched user roles:', gebruiker?.roles);
   if (!gebruiker) {
     throw ServiceError.notFound('Er bestaat geen gebruiker met dit id');
   }
@@ -126,9 +126,22 @@ export const getById = async (id: number): Promise<PublicGebruiker> => {
 
 export const updateById = async (id: number, changes: GebruikerUpdateInput): Promise<PublicGebruiker> => {
   try {
+    const dataToUpdate: any = { ...changes };
+
+    if (changes.wachtwoord) {
+      dataToUpdate.wachtwoord = await hashPassword(changes.wachtwoord);
+    } else {
+      delete dataToUpdate.wachtwoord;
+    }
+
+    // If roles is undefined, do not update roles
+    if (typeof changes.roles === 'undefined') {
+      delete dataToUpdate.roles;
+    }
+
     const user = await prisma.gebruiker.update({
       where: { id },
-      data: changes,
+      data: dataToUpdate,
     });
     return makeExposedUser(user);
   } catch (error) {
@@ -136,10 +149,16 @@ export const updateById = async (id: number, changes: GebruikerUpdateInput): Pro
   }
 };
 
-export const deleteById = async (id: number): Promise<void> => {
+export const deleteById = async (id: number, rolesUser: string[]): Promise<void> => {
+
+  if (!rolesUser.includes('admin')) {
+    throw ServiceError.forbidden('Je hebt geen toestemming om gebruikers te verwijderen');
+  }
+
   try {
     await prisma.gebruiker.delete({ where: { id } });
   } catch (error) {
     throw handleDBError(error);
   }
 };
+
